requirements/main.txt
# The project's dependencies...
cnx-litezip
pyramid
